Refactor: Move fs related functions from utils into their own submodule
