+++ src/book/mdbook.rs
@@ -198,7 +198,7 @@ impl MDBook {
         try!(self.init());
 
         // Clean output directory
-        try!(utils::remove_dir_content(&self.config.get_dest()));
+        try!(utils::fs::remove_dir_content(&self.config.get_dest()));
 
         try!(self.renderer.render(&self));
